crates: add finer-grain protocol crates
Break up the current `protocol` crate into:

* proto-build, which has boilerplate for building protobuf
* proto-gazette, which is prost types + serde support for gazette messages
* proto-flow, which is prost types + serde support for flow messages
* proto-grpc, which is tonic gRPC client & server stubs for services.
  Each client or server is behind a feature flag.

`pbjson` is used to generate protobuf-compliant JSON mappings
for all types.
