+++ crates/proto-build/Cargo.toml
@@ -1,0 +1,5 @@
+[package]
+name = "proto-build"
+version = "0.0.0"
+authors = ["Estuary Technologies, Inc"]
+edition = "2021"
