+++ crates/proto-gazette/build.rs
@@ -1,0 +1,18 @@
+fn main() {
+    let b = proto_build::Boilerplate::create();
+    let proto_build = b.resolve_gazette_targets();
+
+    prost_build::Config::new()
+        .out_dir(&b.src_dir)
+        .file_descriptor_set_path(&b.descriptor_path)
+        .compile_well_known_types()
+        .extern_path(".google.protobuf", "::pbjson_types")
+        .compile_protos(&proto_build, &b.proto_include())
+        .expect("failed to compile protobuf");
+
+    pbjson_build::Builder::new()
+        .register_descriptors(&std::fs::read(b.descriptor_path).expect("read descriptors"))
+        .unwrap()
+        .build(&[".protocol", ".consumer", ".recoverylog"])
+        .expect("building pbjson");
+}
