+++ crates/proto-gazette/src/protocol.rs
@@ -99,12 +99,12 @@ pub mod journal_spec {
         /// Interval of time between refreshes of remote Fragment listings from
         /// configured fragment_stores.
         #[prost(message, optional, tag="4")]
-        pub refresh_interval: ::core::option::Option<::prost_types::Duration>,
+        pub refresh_interval: ::core::option::Option<::pbjson_types::Duration>,
         /// Retention duration for historical Fragments of this Journal within the
         /// Fragment stores. If less than or equal to zero, Fragments are retained
         /// indefinitely.
         #[prost(message, optional, tag="5")]
-        pub retention: ::core::option::Option<::prost_types::Duration>,
+        pub retention: ::core::option::Option<::pbjson_types::Duration>,
         /// Flush interval defines a uniform UTC time segment which, when passed,
         /// will prompt brokers to close and persist a fragment presently being
         /// written.
@@ -119,7 +119,7 @@ pub mod journal_spec {
         ///
         /// See also "gazctl journals fragments --help" for more discussion.
         #[prost(message, optional, tag="6")]
-        pub flush_interval: ::core::option::Option<::prost_types::Duration>,
+        pub flush_interval: ::core::option::Option<::pbjson_types::Duration>,
         /// Path postfix template is a Go template which evaluates to a partial
         /// path under which fragments are persisted to the store. A complete
         /// fragment path is constructed by appending path components from the
@@ -578,7 +578,7 @@ pub struct FragmentsRequest {
     /// SignatureTTL indicates that a temporary signed GET URL should be returned
     /// with each response Fragment, valid for |signatureTTL|.
     #[prost(message, optional, tag="7")]
-    pub signature_ttl: ::core::option::Option<::prost_types::Duration>,
+    pub signature_ttl: ::core::option::Option<::pbjson_types::Duration>,
     /// If do_not_proxy is true, the broker will not proxy the request to another
     /// broker on the client's behalf.
     #[prost(bool, tag="8")]
@@ -749,563 +749,4 @@ pub enum CompressionCodec {
     /// header handling can be subtle and sometimes confusing. It uses the default
     /// suffix ".gzod".
     GzipOffloadDecompression = 5,
-}
-/// Generated client implementations.
-#[cfg(feature = "gaz_broker_client")]
-pub mod journal_client {
-    #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
-    use tonic::codegen::*;
-    /// Journal is the Gazette broker service API for interacting with Journals.
-    #[derive(Debug, Clone)]
-    pub struct JournalClient<T> {
-        inner: tonic::client::Grpc<T>,
-    }
-    impl JournalClient<tonic::transport::Channel> {
-        /// Attempt to create a new client by connecting to a given endpoint.
-        pub async fn connect<D>(dst: D) -> Result<Self, tonic::transport::Error>
-        where
-            D: std::convert::TryInto<tonic::transport::Endpoint>,
-            D::Error: Into<StdError>,
-        {
-            let conn = tonic::transport::Endpoint::new(dst)?.connect().await?;
-            Ok(Self::new(conn))
-        }
-    }
-    impl<T> JournalClient<T>
-    where
-        T: tonic::client::GrpcService<tonic::body::BoxBody>,
-        T::Error: Into<StdError>,
-        T::ResponseBody: Default + Body<Data = Bytes> + Send + 'static,
-        <T::ResponseBody as Body>::Error: Into<StdError> + Send,
-    {
-        pub fn new(inner: T) -> Self {
-            let inner = tonic::client::Grpc::new(inner);
-            Self { inner }
-        }
-        pub fn with_interceptor<F>(
-            inner: T,
-            interceptor: F,
-        ) -> JournalClient<InterceptedService<T, F>>
-        where
-            F: tonic::service::Interceptor,
-            T: tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-                Response = http::Response<
-                    <T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody,
-                >,
-            >,
-            <T as tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-            >>::Error: Into<StdError> + Send + Sync,
-        {
-            JournalClient::new(InterceptedService::new(inner, interceptor))
-        }
-        /// Compress requests with `gzip`.
-        ///
-        /// This requires the server to support it otherwise it might respond with an
-        /// error.
-        #[must_use]
-        pub fn send_gzip(mut self) -> Self {
-            self.inner = self.inner.send_gzip();
-            self
-        }
-        /// Enable decompressing responses with `gzip`.
-        #[must_use]
-        pub fn accept_gzip(mut self) -> Self {
-            self.inner = self.inner.accept_gzip();
-            self
-        }
-        /// List Journals, their JournalSpecs and current Routes.
-        pub async fn list(
-            &mut self,
-            request: impl tonic::IntoRequest<super::ListRequest>,
-        ) -> Result<tonic::Response<super::ListResponse>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static("/protocol.Journal/List");
-            self.inner.unary(request.into_request(), path, codec).await
-        }
-        /// Apply changes to the collection of Journals managed by the brokers.
-        pub async fn apply(
-            &mut self,
-            request: impl tonic::IntoRequest<super::ApplyRequest>,
-        ) -> Result<tonic::Response<super::ApplyResponse>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static("/protocol.Journal/Apply");
-            self.inner.unary(request.into_request(), path, codec).await
-        }
-        /// Read from a specific Journal.
-        pub async fn read(
-            &mut self,
-            request: impl tonic::IntoRequest<super::ReadRequest>,
-        ) -> Result<
-                tonic::Response<tonic::codec::Streaming<super::ReadResponse>>,
-                tonic::Status,
-            > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static("/protocol.Journal/Read");
-            self.inner.server_streaming(request.into_request(), path, codec).await
-        }
-        /// Append content to a specific Journal.
-        pub async fn append(
-            &mut self,
-            request: impl tonic::IntoStreamingRequest<Message = super::AppendRequest>,
-        ) -> Result<tonic::Response<super::AppendResponse>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static("/protocol.Journal/Append");
-            self.inner
-                .client_streaming(request.into_streaming_request(), path, codec)
-                .await
-        }
-        /// Replicate appended content of a Journal. Replicate is used between broker
-        /// peers in the course of processing Append transactions, but is not intended
-        /// for direct use by clients.
-        pub async fn replicate(
-            &mut self,
-            request: impl tonic::IntoStreamingRequest<Message = super::ReplicateRequest>,
-        ) -> Result<
-                tonic::Response<tonic::codec::Streaming<super::ReplicateResponse>>,
-                tonic::Status,
-            > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/protocol.Journal/Replicate",
-            );
-            self.inner.streaming(request.into_streaming_request(), path, codec).await
-        }
-        /// List Fragments of a Journal.
-        pub async fn list_fragments(
-            &mut self,
-            request: impl tonic::IntoRequest<super::FragmentsRequest>,
-        ) -> Result<tonic::Response<super::FragmentsResponse>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/protocol.Journal/ListFragments",
-            );
-            self.inner.unary(request.into_request(), path, codec).await
-        }
-    }
-}
-/// Generated server implementations.
-#[cfg(feature = "gaz_broker_server")]
-pub mod journal_server {
-    #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
-    use tonic::codegen::*;
-    ///Generated trait containing gRPC methods that should be implemented for use with JournalServer.
-    #[async_trait]
-    pub trait Journal: Send + Sync + 'static {
-        /// List Journals, their JournalSpecs and current Routes.
-        async fn list(
-            &self,
-            request: tonic::Request<super::ListRequest>,
-        ) -> Result<tonic::Response<super::ListResponse>, tonic::Status>;
-        /// Apply changes to the collection of Journals managed by the brokers.
-        async fn apply(
-            &self,
-            request: tonic::Request<super::ApplyRequest>,
-        ) -> Result<tonic::Response<super::ApplyResponse>, tonic::Status>;
-        ///Server streaming response type for the Read method.
-        type ReadStream: futures_core::Stream<
-                Item = Result<super::ReadResponse, tonic::Status>,
-            >
-            + Send
-            + 'static;
-        /// Read from a specific Journal.
-        async fn read(
-            &self,
-            request: tonic::Request<super::ReadRequest>,
-        ) -> Result<tonic::Response<Self::ReadStream>, tonic::Status>;
-        /// Append content to a specific Journal.
-        async fn append(
-            &self,
-            request: tonic::Request<tonic::Streaming<super::AppendRequest>>,
-        ) -> Result<tonic::Response<super::AppendResponse>, tonic::Status>;
-        ///Server streaming response type for the Replicate method.
-        type ReplicateStream: futures_core::Stream<
-                Item = Result<super::ReplicateResponse, tonic::Status>,
-            >
-            + Send
-            + 'static;
-        /// Replicate appended content of a Journal. Replicate is used between broker
-        /// peers in the course of processing Append transactions, but is not intended
-        /// for direct use by clients.
-        async fn replicate(
-            &self,
-            request: tonic::Request<tonic::Streaming<super::ReplicateRequest>>,
-        ) -> Result<tonic::Response<Self::ReplicateStream>, tonic::Status>;
-        /// List Fragments of a Journal.
-        async fn list_fragments(
-            &self,
-            request: tonic::Request<super::FragmentsRequest>,
-        ) -> Result<tonic::Response<super::FragmentsResponse>, tonic::Status>;
-    }
-    /// Journal is the Gazette broker service API for interacting with Journals.
-    #[derive(Debug)]
-    pub struct JournalServer<T: Journal> {
-        inner: _Inner<T>,
-        accept_compression_encodings: (),
-        send_compression_encodings: (),
-    }
-    struct _Inner<T>(Arc<T>);
-    impl<T: Journal> JournalServer<T> {
-        pub fn new(inner: T) -> Self {
-            Self::from_arc(Arc::new(inner))
-        }
-        pub fn from_arc(inner: Arc<T>) -> Self {
-            let inner = _Inner(inner);
-            Self {
-                inner,
-                accept_compression_encodings: Default::default(),
-                send_compression_encodings: Default::default(),
-            }
-        }
-        pub fn with_interceptor<F>(
-            inner: T,
-            interceptor: F,
-        ) -> InterceptedService<Self, F>
-        where
-            F: tonic::service::Interceptor,
-        {
-            InterceptedService::new(Self::new(inner), interceptor)
-        }
-    }
-    impl<T, B> tonic::codegen::Service<http::Request<B>> for JournalServer<T>
-    where
-        T: Journal,
-        B: Body + Send + 'static,
-        B::Error: Into<StdError> + Send + 'static,
-    {
-        type Response = http::Response<tonic::body::BoxBody>;
-        type Error = std::convert::Infallible;
-        type Future = BoxFuture<Self::Response, Self::Error>;
-        fn poll_ready(
-            &mut self,
-            _cx: &mut Context<'_>,
-        ) -> Poll<Result<(), Self::Error>> {
-            Poll::Ready(Ok(()))
-        }
-        fn call(&mut self, req: http::Request<B>) -> Self::Future {
-            let inner = self.inner.clone();
-            match req.uri().path() {
-                "/protocol.Journal/List" => {
-                    #[allow(non_camel_case_types)]
-                    struct ListSvc<T: Journal>(pub Arc<T>);
-                    impl<T: Journal> tonic::server::UnaryService<super::ListRequest>
-                    for ListSvc<T> {
-                        type Response = super::ListResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::ListRequest>,
-                        ) -> Self::Future {
-                            let inner = self.0.clone();
-                            let fut = async move { (*inner).list(request).await };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = ListSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/protocol.Journal/Apply" => {
-                    #[allow(non_camel_case_types)]
-                    struct ApplySvc<T: Journal>(pub Arc<T>);
-                    impl<T: Journal> tonic::server::UnaryService<super::ApplyRequest>
-                    for ApplySvc<T> {
-                        type Response = super::ApplyResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::ApplyRequest>,
-                        ) -> Self::Future {
-                            let inner = self.0.clone();
-                            let fut = async move { (*inner).apply(request).await };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = ApplySvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/protocol.Journal/Read" => {
-                    #[allow(non_camel_case_types)]
-                    struct ReadSvc<T: Journal>(pub Arc<T>);
-                    impl<
-                        T: Journal,
-                    > tonic::server::ServerStreamingService<super::ReadRequest>
-                    for ReadSvc<T> {
-                        type Response = super::ReadResponse;
-                        type ResponseStream = T::ReadStream;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::ResponseStream>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::ReadRequest>,
-                        ) -> Self::Future {
-                            let inner = self.0.clone();
-                            let fut = async move { (*inner).read(request).await };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = ReadSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            );
-                        let res = grpc.server_streaming(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/protocol.Journal/Append" => {
-                    #[allow(non_camel_case_types)]
-                    struct AppendSvc<T: Journal>(pub Arc<T>);
-                    impl<
-                        T: Journal,
-                    > tonic::server::ClientStreamingService<super::AppendRequest>
-                    for AppendSvc<T> {
-                        type Response = super::AppendResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<
-                                tonic::Streaming<super::AppendRequest>,
-                            >,
-                        ) -> Self::Future {
-                            let inner = self.0.clone();
-                            let fut = async move { (*inner).append(request).await };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = AppendSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            );
-                        let res = grpc.client_streaming(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/protocol.Journal/Replicate" => {
-                    #[allow(non_camel_case_types)]
-                    struct ReplicateSvc<T: Journal>(pub Arc<T>);
-                    impl<
-                        T: Journal,
-                    > tonic::server::StreamingService<super::ReplicateRequest>
-                    for ReplicateSvc<T> {
-                        type Response = super::ReplicateResponse;
-                        type ResponseStream = T::ReplicateStream;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::ResponseStream>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<
-                                tonic::Streaming<super::ReplicateRequest>,
-                            >,
-                        ) -> Self::Future {
-                            let inner = self.0.clone();
-                            let fut = async move { (*inner).replicate(request).await };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = ReplicateSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            );
-                        let res = grpc.streaming(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/protocol.Journal/ListFragments" => {
-                    #[allow(non_camel_case_types)]
-                    struct ListFragmentsSvc<T: Journal>(pub Arc<T>);
-                    impl<T: Journal> tonic::server::UnaryService<super::FragmentsRequest>
-                    for ListFragmentsSvc<T> {
-                        type Response = super::FragmentsResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::FragmentsRequest>,
-                        ) -> Self::Future {
-                            let inner = self.0.clone();
-                            let fut = async move {
-                                (*inner).list_fragments(request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = ListFragmentsSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                _ => {
-                    Box::pin(async move {
-                        Ok(
-                            http::Response::builder()
-                                .status(200)
-                                .header("grpc-status", "12")
-                                .header("content-type", "application/grpc")
-                                .body(empty_body())
-                                .unwrap(),
-                        )
-                    })
-                }
-            }
-        }
-    }
-    impl<T: Journal> Clone for JournalServer<T> {
-        fn clone(&self) -> Self {
-            let inner = self.inner.clone();
-            Self {
-                inner,
-                accept_compression_encodings: self.accept_compression_encodings,
-                send_compression_encodings: self.send_compression_encodings,
-            }
-        }
-    }
-    impl<T: Journal> Clone for _Inner<T> {
-        fn clone(&self) -> Self {
-            Self(self.0.clone())
-        }
-    }
-    impl<T: std::fmt::Debug> std::fmt::Debug for _Inner<T> {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            write!(f, "{:?}", self.0)
-        }
-    }
-    impl<T: Journal> tonic::transport::NamedService for JournalServer<T> {
-        const NAME: &'static str = "protocol.Journal";
-    }
 }
