+++ crates/proto-gazette/src/recoverylog.rs
