+++ crates/proto-grpc/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod capture;
+pub mod consumer;
+pub mod flow;
+pub mod materialize;
+mod protocol;
+
+// The `protocol` package is publicly exported as `broker`.
+pub mod broker {
+    pub use crate::protocol::*;
+}
