+++ crates/protocol/Cargo.toml
@@ -5,49 +5,16 @@ authors = ["Estuary Technologies, Inc"]
 edition = "2018"
 
 [dependencies]
+
+proto-gazette = { path = "../proto-gazette" }
+proto-flow = { path = "../proto-flow" }
+
 anyhow = "*"
 bytes = "*"
-humantime-serde = "*"
 prost = "*"
-prost-types = "*"
-tonic = {version = "*", optional = true}
-tonic-types = {version = "*", optional = true}
 serde = "*"
 tracing = "*"
 
 [dev-dependencies]
 insta = "*"
 serde_json = "*"
-
-[build-dependencies]
-prost-build = "*"
-tonic-build = "*"
-
-
-[features]
-default = []
-# The test-support feature enables additional functionality for use during tests. This feature is
-# intended to only be used from "dev-dependencies"
-test-support = []
-
-# The grpc feature is required for any client/server features.
-grpc = ["tonic", "tonic-types"]
-
-# The features below correspond to the various services defined in the protobuf files.
-# The conditional compilation attributes are setup in `build.rs`, so this list of features matches
-# the ones used there.
-materialize_client = ["grpc"]
-materialize_server = ["grpc"]
-
-capture_client = ["grpc"]
-capture_server = ["grpc"]
-
-flow_client = ["grpc"]
-flow_server = ["grpc"]
-
-gaz_consumer_client = ["grpc"]
-gaz_consumer_server = ["grpc"]
-
-gaz_broker_client = ["grpc"]
-gaz_broker_server = ["grpc"]
-
